Add VOLK_NAMESPACE configuration to use C++ namespaces
In some cases the use of global vk* symbols is inconvenient, as it can
conflict with vk* functions defined by the loader; generally speaking
this should not be required, but some dynamic library cases are
difficult to support cleanly.

This problem can be solved when C++ is available: by building volk.h/c
as C++ and wrapping functions in volk namespace, which is automatically
used, the compiler can still find the functions so no change in source
code is necessary, but the global symbols get C++ mangling that never
overlaps with vk*.

This is not designed as a main mode of operation, but it can be useful
when the default setup doesn't fully work.